Feat: 새로운 아이템 유형을 기반으로한 FertilizerItem 클래스의 작성

+++ Assets/Scripts/Runtime/Item/ChargeInfo.cs
@@ -7,9 +7,9 @@ using UnityEngine;
 public struct ChargeInfo
 {
     public float ChargeTime;
-    public sbyte MaxChargeCount;
+    public int MaxChargeCount;
 
-    public ChargeInfo(float chargeTime, sbyte maxChargeCount)
+    public ChargeInfo(float chargeTime, int maxChargeCount)
     {
         ChargeTime = chargeTime;
         MaxChargeCount = maxChargeCount;
